fixup! rename SearchIncidents_v2 -> SearchIncidentsv2
