+++ Playbooks/playbook-Send_Investigation_Summary_Reports.yml
@@ -40,7 +40,7 @@ tasks:
       version: -1
       name: Find Closed Incidents
       description: Searches Demisto incidents to identify closed incidents.
-      scriptName: SearchIncidentsv2
+      scriptName: SearchIncidents
       type: regular
       iscommand: false
       brand: ""
