+++ Playbooks/playbook-Send_Investigation_Summary_Reports_CHANGELOG.md
@@ -1,2 +1,1 @@
 ## [Unreleased]
-Replace searchIncidents with SearchIncidentsv2
