+++ TestPlaybooks/playbook-Send_Investigation_Summary_Reports_Test.yml
@@ -295,7 +295,7 @@ tasks:
       version: -1
       name: Get Incident
       description: Searches Demisto incidents
-      scriptName: SearchIncidentsv2
+      scriptName: SearchIncidents
       type: regular
       iscommand: false
       brand: ""
